Finds midpoints and bands for the within and between cycles.

Finds midpoints and bands for the within and between cycles. This the second of two functions that needs to be called to produce WATS Plots. AugmentZZZ is the first.

Usage

AnnotateData(
  dsLinear,
  dvName,
  centerFunction,
  spreadFunction,
  cycleTallyName = "CycleTally",
  stageIDName = "StageID",
  stageProgressName = "StageProgress",
  proportionThroughCycleName = "ProportionThroughCycle",
  proportionIDName = "ProportionID",
  terminalPointInCycleName = "TerminalPointInCycle"
)

Arguments

dsLinear

The data.frame to containing the detailed data.

dvName

The name of the dependent/criterion variable.

centerFunction

A function to calculate the center of a subsample.

spreadFunction

A function to calculate the bands of a subsample.

cycleTallyName

The variable name indicating how many cycles have been completed.

stageIDName

The variable name indicating the stage. In a typical interrupted time series, these values are 1 before the interruption and 2 after.

stageProgressName

The variable name indicating the stage in a decimal form. This is mostly for internal uses.

proportionThroughCycleName

The variable name indicating how far the point is through a cycle. For example, 0 degrees would be 0, 180 degrees would be 0.5, 359 degrees would be 0.9972, and 360 degrees would be 0.

proportionIDName

The variable name indicating the ordinal position through a cycle.

terminalPointInCycleName

The variable name indicating the last point within a given cycle.

Value

Returns a data.frame with additional variables <>.

Examples

library(Wats)
dsLinear <- CountyMonthBirthRate2005Version
dsLinear <- dsLinear[dsLinear$CountyName=="oklahoma", ]
dsLinear <- AugmentYearDataWithMonthResolution(dsLinear=dsLinear, dateName="Date")

hSpread <- function( scores ) { return( quantile(x=scores, probs=c(.25, .75)) ) }
portfolio <- AnnotateData(
  dsLinear = dsLinear,
  dvName = "BirthRate",
  centerFunction = median,
  spreadFunction = hSpread
)

head(portfolio$dsStageCycle)
#> # A tibble: 6 × 6
#>   StageID ProportionID ProportionThroughCycle PositionLower PositionCe…¹ Posit…²
#>     <int>        <dbl>                  <dbl>         <dbl>        <dbl>   <dbl>
#> 1       1            1                 0.0417          66.8         70.1    70.6
#> 2       1            2                 0.125           68.9         69.5    71.7
#> 3       1            3                 0.208           66.1         69.7    69.9
#> 4       1            4                 0.292           67.7         69.3    71.6
#> 5       1            5                 0.375           67.7         68.6    70.6
#> 6       1            6                 0.458           70.9         71.1    72.9
#> # … with abbreviated variable names ¹​PositionCenter, ²​PositionUpper
head(portfolio$dsLinear)
#> # A tibble: 6 × 21
#>    Fips CountyN…¹  Year Month Fecun…² Birth…³ Date       DaysI…⁴ DaysI…⁵ StageID
#>   <int> <chr>     <int> <int>   <dbl>   <int> <date>       <int>   <int>   <int>
#> 1 40109 oklahoma   1990     1 143192      853 1990-01-15      31     365       1
#> 2 40109 oklahoma   1990     2 143279.     758 1990-02-15      28     365       1
#> 3 40109 oklahoma   1990     3 143365.     886 1990-03-15      31     365       1
#> 4 40109 oklahoma   1990     4 143452      871 1990-04-15      30     365       1
#> 5 40109 oklahoma   1990     5 143539.     822 1990-05-15      31     365       1
#> 6 40109 oklahoma   1990     6 143625.     834 1990-06-15      30     365       1
#> # … with 11 more variables: BirthRateMonthly <dbl>, BirthRate <dbl>,
#> #   CycleTally <dbl>, ProportionThroughCycle <dbl>, ProportionID <dbl>,
#> #   StartingPointInCycle <lgl>, TerminalPointInCycle <lgl>,
#> #   StageProgress <dbl>, RollingLower <dbl>, RollingCenter <dbl>,
#> #   RollingUpper <dbl>, and abbreviated variable names ¹​CountyName,
#> #   ²​FecundPopulation, ³​BirthCount, ⁴​DaysInMonth, ⁵​DaysInYear
head(portfolio$dsPeriodic)
#> # A tibble: 6 × 9
#>   Date       StageIDTime Propo…¹ Stage…² Stage…³ Propo…⁴ Posit…⁵ Posit…⁶ Posit…⁷
#>   <date>           <int>   <dbl>   <dbl>   <int>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 1990-01-15           1       1     1         1  0.0417    66.8    70.1    70.6
#> 2 1990-01-15           1       1     1         2  0.0417    70.8    73.5    73.6
#> 3 1990-02-15           1       2     1.5       1  0.125     68.9    69.5    71.7
#> 4 1990-02-15           1       2     1.5       2  0.125     69.3    72.0    74.1
#> 5 1990-03-15           1       3     1.5       1  0.208     66.1    69.7    69.9
#> 6 1990-03-15           1       3     1.5       2  0.208     70.7    71.1    72.5
#> # … with abbreviated variable names ¹​ProportionID, ²​StageProgress,
#> #   ³​StageIDBand, ⁴​ProportionThroughCycle, ⁵​PositionLower, ⁶​PositionCenter,
#> #   ⁷​PositionUpper