Wrap Around Time Series graphics

Wrap-around Time Series (WATS) Plots for Interrupted Time Series Designs with Seasonal Patterns

Note

The release version is available through CRAN by running install.packages('Wats'). The most recent development version is available through GitHub by running remotes::install_github("OuhscBbmc/Wats"). (make sure remotes is already installed). If you're having trouble with the package, please install the development version. If this doesn't solve your problem, please create an issue, or email Will.

Funding Continued development and maintenance of this package supported by the Oklahoma Shared Clinical and Translational Resources (OSCTR, U54GM104938) with an Institutional Development Award (IDeA) from NIGMS. The content is solely the responsibility of the authors and does not necessarily represent the official views of the National Institutes of Health.

References

Rodgers, J.L., Beasley, W.H., and Schuelke, M. (2014). Graphical Data Analysis on the Circle: Wrap-around Time Series Plots for (Interrupted) Time Series Designs. Multivariate Behavioral Research.

Rodgers, J.L., St. John, C. A. & Coleman R. (2005). Did Fertility Go Up after the Oklahoma City Bombing? An Analysis of Births in Metropolitan Counties in Oklahoma, 1990-1999. Demography, 42, 675-692.

Author

William Howard Beasley –Associate Professor of Research, University of Oklahoma Health Sciences Center, Dept of Pediatrics, Biomedical and Behavioral Methodology Core (BBMC)

Joseph Lee Rodgers –Director, Quantitative Methods, Department of Psychology and Human Development, Peabody College, 230 Appleton Pl #552, Hobbs 202C, Vanderbilt University, Nashville, TN 37203.

Matthew Schuelke –Senior Statistician, Office of the Vice President for Research Saint Louis University