Finds midpoints and bands for the within and between cycles.

Finds midpoints and bands for the within and between cycles. This the second of two functions that needs to be called to produce WATS Plots. AugmentZZZ is the first.

Usage

AnnotateData(
  dsLinear,
  dvName,
  centerFunction,
  spreadFunction,
  cycleTallyName = "CycleTally",
  stageIDName = "StageID",
  stageProgressName = "StageProgress",
  proportionThroughCycleName = "ProportionThroughCycle",
  proportionIDName = "ProportionID",
  terminalPointInCycleName = "TerminalPointInCycle"
)

Arguments

dsLinear

The data.frame to containing the detailed data.

dvName

The name of the dependent/criterion variable.

centerFunction

A function to calculate the center of a subsample.

spreadFunction

A function to calculate the bands of a subsample.

cycleTallyName

The variable name indicating how many cycles have been completed.

stageIDName

The variable name indicating the stage. In a typical interrupted time series, these values are 1 before the interruption and 2 after.

stageProgressName

The variable name indicating the stage in a decimal form. This is mostly for internal uses.

proportionThroughCycleName

The variable name indicating how far the point is through a cycle. For example, 0 degrees would be 0, 180 degrees would be 0.5, 359 degrees would be 0.9972, and 360 degrees would be 0.

proportionIDName

The variable name indicating the ordinal position through a cycle.

terminalPointInCycleName

The variable name indicating the last point within a given cycle.

Value

Returns a data.frame with additional variables <>.

Examples

library(Wats)
dsLinear <- CountyMonthBirthRate2005Version
dsLinear <- dsLinear[dsLinear$CountyName=="oklahoma", ]
dsLinear <- AugmentYearDataWithMonthResolution(dsLinear=dsLinear, dateName="Date")

hSpread <- function( scores ) { return( quantile(x=scores, probs=c(.25, .75)) ) }
portfolio <- AnnotateData(
  dsLinear = dsLinear,
  dvName = "BirthRate",
  centerFunction = median,
  spreadFunction = hSpread
)

head(portfolio$dsStageCycle)
#>   StageID ProportionID ProportionThroughCycle PositionLower PositionCenter
#> 1       1            1             0.04166667      66.82005       70.13930
#> 2       1            2             0.12500000      68.93747       69.53782
#> 3       1            3             0.20833333      66.13076       69.70012
#> 4       1            4             0.29166667      67.71910       69.30147
#> 5       1            5             0.37500000      67.65719       68.58747
#> 6       1            6             0.45833333      70.88414       71.13025
#>   PositionUpper
#> 1      70.55044
#> 2      71.72977
#> 3      69.91971
#> 4      71.56912
#> 5      70.55044
#> 6      72.88710
head(portfolio$dsLinear)
#>      Fips CountyName Year Month FecundPopulation BirthCount       Date
#> 721 40109   oklahoma 1990     1         143192.0        853 1990-01-15
#> 722 40109   oklahoma 1990     2         143278.7        758 1990-02-15
#> 723 40109   oklahoma 1990     3         143365.3        886 1990-03-15
#> 724 40109   oklahoma 1990     4         143452.0        871 1990-04-15
#> 725 40109   oklahoma 1990     5         143538.7        822 1990-05-15
#> 726 40109   oklahoma 1990     6         143625.3        834 1990-06-15
#>     DaysInMonth DaysInYear StageID BirthRateMonthly BirthRate CycleTally
#> 721          31        365       1         5.957037  70.13930          0
#> 722          28        365       1         5.293592  69.00575          0
#> 723          31        365       1         6.187497  72.85278          0
#> 724          30        365       1         6.082742  74.00669          0
#> 725          31        365       1         5.740544  67.59028          0
#> 726          30        365       1         5.824348  70.86290          0
#>     ProportionThroughCycle ProportionID StartingPointInCycle
#> 721             0.04166667            1                 TRUE
#> 722             0.12500000            2                FALSE
#> 723             0.20833333            3                FALSE
#> 724             0.29166667            4                FALSE
#> 725             0.37500000            5                FALSE
#> 726             0.45833333            6                FALSE
#>     TerminalPointInCycle StageProgress RollingLower RollingCenter RollingUpper
#> 721                FALSE           1.0           NA            NA           NA
#> 722                FALSE           1.5           NA            NA           NA
#> 723                FALSE           1.5           NA            NA           NA
#> 724                FALSE           1.5           NA            NA           NA
#> 725                FALSE           1.5           NA            NA           NA
#> 726                FALSE           1.5           NA            NA           NA
head(portfolio$dsPeriodic)
#>    ProportionID       Date StageIDTime StageProgress StageIDBand
#> 1             1 1990-01-15           1           1.0           1
#> 2             1 1990-01-15           1           1.0           2
#> 21            2 1990-02-15           1           1.5           1
#> 22            2 1990-02-15           1           1.5           2
#> 43            3 1990-03-15           1           1.5           1
#> 44            3 1990-03-15           1           1.5           2
#>    ProportionThroughCycle PositionLower PositionCenter PositionUpper
#> 1              0.04166667      66.82005       70.13930      70.55044
#> 2              0.04166667      70.75600       73.51059      73.59282
#> 21             0.12500000      68.93747       69.53782      71.72977
#> 22             0.12500000      69.34846       71.96444      74.14932
#> 43             0.20833333      66.13076       69.70012      69.91971
#> 44             0.20833333      70.70149       71.12602      72.54443